dev: add test for header hash (#695)
